backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
